save it!
So far everything is working, the more types we put in there the more
clearer everything becomes. Cameron provided the idea of attaching a
method to the Mix class so that we may call the base methods, its just
a simple wrapper so that we don’t have to type prototype.